requirements.txt
Flask
flask_cors
pycryptodome
